requirements.txt
apache-airflow[celery]==3.0.4
black
